forge: move HealthCheckError to interface module
